requirements.txt
google-cloud-pubsub
apache-beam[gcp,test]==2.42.0
google-cloud-storage==2.14.0